+++ packages/editor/js/extensions/libs/block-states/components/states-manager.js
@@ -26,7 +26,6 @@ import {
  * Internal dependencies
  */
 import ItemBody from './item-body';
-import defaultStates from '../states';
 import ItemHeader from './item-header';
 import ItemOpener from './item-opener';
 import {
@@ -71,11 +70,12 @@ const StatesManager: ComponentType<any> = memo(
 		} = dispatch('blockera/extensions') || {};
 		const { getBlockStates, getActiveMasterState, getActiveInnerState } =
 			select('blockera/extensions');
-		const { getBreakpoints } = select('blockera/editor');
+		const { getStates, getBreakpoints } = select('blockera/editor');
 		const savedBlockStates = getBlockStates(
 			block?.clientId,
 			!isMasterBlockStates(id) ? currentBlock : block?.blockName
 		);
+		const defaultStates = getStates();
 		const clonedSavedStates = { ...states };
 
 		if (isEmpty(states)) {
@@ -293,7 +293,8 @@ const StatesManager: ComponentType<any> = memo(
 					...defaultRepeaterItemValue,
 					...getStateInfo(
 						// If deletedItems has items try to add first index of that else add suitable items for statesCount value.
-						deletedItems?.length ? deletedItems[0] : statesCount
+						deletedItems?.length ? deletedItems[0] : statesCount,
+						defaultStates
 					),
 					display: true,
 				};
@@ -320,17 +321,21 @@ const StatesManager: ComponentType<any> = memo(
 		);
 		const handleOnChange = useCallback(
 			(newValue: Object) =>
-				onChangeBlockStates(newValue, {
-					states,
-					onChange,
-					currentState,
-					currentBlock,
-					valueCleanup,
-					getStateInfo,
-					getBlockStates,
-					currentInnerBlockState,
-					isMasterBlockStates: isMasterBlockStates(id),
-				}),
+				onChangeBlockStates(
+					newValue,
+					{
+						states,
+						onChange,
+						currentState,
+						currentBlock,
+						valueCleanup,
+						getStateInfo,
+						getBlockStates,
+						currentInnerBlockState,
+						isMasterBlockStates: isMasterBlockStates(id),
+					},
+					defaultStates
+				),
 			// eslint-disable-next-line
 			[currentBlock, currentInnerBlockState, currentState, id, states]
 		);
@@ -384,7 +389,15 @@ const StatesManager: ComponentType<any> = memo(
 							overrideItem,
 							repeaterItemHeader: ItemHeader,
 							repeaterItemOpener: ItemOpener,
-							repeaterItemChildren: ItemBody,
+							repeaterItemChildren: (
+								props: Object
+							): MixedElement => {
+								return (
+									<ItemBody
+										{...{ ...props, states: defaultStates }}
+									/>
+								);
+							},
 						}}
 						defaultValue={states}
 						addNewButtonLabel={__('Add New State', 'blockera')}