+++ packages/editor/js/extensions/libs/block-states/test/default-breakpoints.spec.js
@@ -1,7 +1,6 @@
 /**
  * Internal dependencies
  */
-import { getBreakpointInfo, getStateInfo } from '../helpers';
 import { default as defaultBreakpoints } from '../default-breakpoints';
 
 describe('Default breakpoints settings', () => {